Improve ability to parse strings and numbers:
* Multi digit numbers are parseable now
* Strings (without escape codes) are parseable
* Add sanity checks for more general edge cases

I am adding a bunch of validation logic which makes it easier to build and
test that the parser is working correctly, but this extra validation logic
takes up space. In the future I plan to make the error validation logic
optional. This would allow people who trust their data to build the library
with error validation srripped out.