Cleanup `test.c`:
Instead of creating a static struct for each test, there is now just one struct
shared across all tests. This is fine because `josh_extract` calls memset()
before doing anything, so the state is reset. The reason for doing this is
because I am planning on writing a custom memory allocator for Josh, and having
lots of function-level structs causes the stack to get too big.