Validate that string used in dot notation is valid JavaScript identifier:
Basically allow only [A-Za-z0-9_] chars when using the dot notation. Before any
and all characters would be allowed, which needed to be fixed before I could
add nested levels of parsing. Although 0-9 is technically not allowed at the
start of JS identifiers, I opted to keep it in because it makes the code
simpler, and it would be an unnecessary restriction in an effort to make things
"compliant" for no other reason than to be compliant.