Add ability to extract from nested JSON:
Also some bug fixes:

* The max key depth is now checked (before it would just write into arbitrary
memory).

* Key dot notation is allowed to be used in conjunction with other keys now.
The string ".a.b.c" was causing errors when it hit the second dot. In addition,
we now have to make a copy of the strings used in the keys, because JSON keys
are able to contain dots (such as "a.b.c"). This is indistinguishable from a
3-level nested key "a.b.c". Some solutions would be storing the length of the
string key, or creating a copy of the string. I am going to opt for the copy
version, simply because I need to write a custom memory allocator anyways (for
storing the eventual JSON tree), so this gives me a good excuse/starting ground
for working on it.